requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
